Creates tests for API endpoints, reconfigures server start
Note: supertest and Jest did not play well with handling the server connection, which is
why the test command was changed to "jest --forceExit --detectOpenHandles"; it's a well-
documented bug online. I changed the server start to exist in server/start.js to avoid one
"open handle" according to Jest.